pyproject.toml
[project]
name = "python-ml-project-template"
version = "0.1.0"
description = "A Python Package Template"
readme = "README.md"
requires-python = ">=3.6"
license = { file = "LICENSE.txt" }
authors = [{ email = "baeisner@andrew.cmu.edu", name = "Ben Eisner" }]
dependencies = [
  "hydra-core == 1.3.2",
  "lightning == 2.0.3",
  "omegaconf == 2.3.0",
  "pandas",
  "torch == 2.0.1",        # CUDA 11.8
  "torchmetrics",
  "torchvision == 0.15.2", # CUDA 11.8
  "wandb == 0.15.4",
]

[build-system]
requires = ["setuptools >= 62.3.2", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
develop = [
  "autoflake == 2.1.1",
  "black == 23.3.0",
  "isort == 5.12.0",
  "mypy == 1.3.0",
  "pandas-stubs == 2.0.2.230605",
  "pylint == 2.17.4",
  "pytest == 7.3.2",
  "pre-commit == 3.3.3",
]
notebooks = ["jupyter"]
build_docs = ["mkdocs-material", "mkdocstrings[python]"]

# This is required to allow us to have notebooks/ at the top level.
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
python_ml_project_template = ["py.typed"]

[tool.isort]
profile = "black"
known_third_party = "wandb"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["torchvision.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = "tests"

[tool.pylint]
known-third-party = "wandb"

[tool.pylint.TYPECHECK]
generated-members = 'torch.*'
